Add Sphinx documentation set-up
This commit adds a Sphinx documentation toolchain to the project
repository. Sphinx is the most common documentation toolchain in Python.
It works with both Markdown and ReStructuredText for documentation
format.

Additionally, it is also trivial to set up free documentation hosting on
ReadTheDocs by using a Sphinx docs tool. This way, our documentation
will be online and hosted ASAP. Changes will take appear almost as soon
as they are made.

Some simple examples exist for now, but we can add more serious examples
in the coming future.

Signed-off-by: Justin W. Flory <git@jwf.io>