Add Dockerfile with docker-compose orchestration, edit grasa_event_locator settings (#11)
* Adding mysql container

* adding docker-compose.yml and adding ignore rule for mysql/

* Add --system flag to pipenv install step

This commit adds the `--system` flag to `pipenv install`. This ensures packages are installed using system paths instead of installing into a virtual environment.

* Use a statically versioned database in docker-compose

This commit changes the database to use the `mariadb:10.1` image currently available for installation via the IUS repositories on CentOS 7. This helps us bring closer parity between local development environment and the final production environment.

* Dockerfile: Fix flag typo

This commit fixes a hyphen typo in the Dockerfile.

* Update Dockerfile

Fixed white space and postgres and --system flag

* Volumes: Set volume path to cache directory

This commit sets the local volume path to a location in the user's home directory instead of a folder at root. The user running the script may not have a `/grasa` directory at root.

* Revert changes to production.py

Limit hard-coded options to the development settings file in `development.py`.

* Update production.py

* Fix conflicting changes

The `--system` flag was duplicated and the `apt-get` steps required manual intervention. This fixes them and condenses the `RUN` steps to a single layer.

* Revert addition to gitignore

Since the MySQL database will be stored somewhere else (i.e. not in the git repo), this path is not required in the `gitignore`.