docs: Add "Features" page
This commit adds a new "Features" page to our project documentation. The
page is similar to what we already wrote in the Phase Gate 1, but it
adds specific technology decisions and a short explanation of how we
plan to accomplish each feature.

Instead of keeping this in our Google Drive, I figured to make this a
public file as part of our development documentation. Once we get closer
to delivering this to the client, we can figure out how to separate the
docs into different audiences.

Signed-off-by: Justin W. Flory <git@jwf.io>