Logged In User
A little experiment, testing whether it is possible to get the Organization Name to appear in the header when a user is logged in. The new code in views.py writes the org_name to the p_test variable, the header information currently lacks a variable to put that p_test in, to be implemented some time soon.