Use WORKDIR path to start Django application
This commit changes the path used to start the Django application to use the default path specified in `WORKDIR`. When the `WORKDIR` is set, this is the relative path used for all commands in a Dockerfile. So if for some reason, later on you need to change the path, you only need to change the `WORKDIR` instead of in several files.