Update Dockerfile build instructions, add docker-compose file, add dev env docs (#20)
* adding script to start the app in a container for development

* temp fix

add a temp fix to mysqlclient not properly installing, and the django server not starting properly.

* changing port 80 to 80000

* adding ingrid as allowed host in development

* docs: Reformat / expand on dev environment doc

This commit  builds on @nlMeminger's initial steps and adds more Markdown syntax to make it easier to read. I also added in some other steps to make it easier for other team members to understand these instructions depending on their operating system.

* Update Dockerfile

* Delete up.sh

* Remove port 443 and change port 8000 to 80

* Use WORKDIR path to start Django application

This commit changes the path used to start the Django application to use the default path specified in `WORKDIR`. When the `WORKDIR` is set, this is the relative path used for all commands in a Dockerfile. So if for some reason, later on you need to change the path, you only need to change the `WORKDIR` instead of in several files.