Front End Merge
The following commit merges Django code with the most recent Front-End changes. This is composed of what was in Master on the morning of October 3rd, and an updated stylesheet and createEvent.php page provided by Lauren via Slack a little later in the day. Account creation and event creation work, the same state we were in like in class.