Database work (#26)
* Initial database models 

These are the very basic models we will be using without having resources or search analytics. May need to change user model once I dig further into authentication.

* Query Methods Added

This commit adds most of the query methods I could think of for future functionality, such as User.get_all_pending(). User model and some of it's methods are likely to change when I get the custom authentication model figured out.

* User approval functionality

Took what Harrison had from yesterday and used it to finalize our authentication. Also added some connectivity from admin.php to the backend, so now user will populate from the database and can be approved or denied.

* Update header.php

Will be using the database branch for new commits so that we are all on one branch.

This commit reimplements a feature: if an admin is logged in, the provider header link will not appear. If a provider is logged in, the admin header link will disappear.

* Logged In User

For quick reference, the additional lines, make it so that the organization name appears in the header if a provider is logged in. Nothing appears if an admin is logged in.

* Redirect by user type

This commit makes it so that if the logged in user is an admin, they will be redirected to the admin page post login.  Providers will be sent to the provider page.

* Admin/User Provider page

If an admin tries to go to the provider page by manually typing in the URL, force them to redirect to index.php. Providers continue to the provider page like normal. Also replaced render with httpresponseredirect so url becomes the intended url, rather than the login url.

* Update admin.php

Adding portion that indicates what email the user is using so that the admin can contact if nessasary. Also included in a mailto link for easy access to an email program Will need to discuss whether we should go ahead with this.

* Finished hooking up admin page to backend

I finished hooking up admin page to backend, with approvals and denials for events and edits, as well as having them pull their data from the DB. I also hooked the individual event pages up to the backend, for the most part. Some work will still need to be done on that page

* Create Event / Views

Begninning of the implementation of the event creation page.

* Added extra fields to the program model

Added some extra fields to accomodate the addition fields in the registration page, will figure out what to make those specifically.

Changed createEvent to write it's values to POST. Still a bit buggy.

Wrote the function to write new program to database.

* Switched html src

Switched html src tag to django static file

* Fixing docker-compose.yml to match master

* changing dockerfile to match master

* Added in ability to see approved events on provider page.

* Commit before presentation

Some imports were missing, they were put back. Just change the database credentials.

* Front End Merge

The following commit merges Django code with the most recent Front-End changes. This is composed of what was in Master on the morning of October 3rd, and an updated stylesheet and createEvent.php page provided by Lauren via Slack a little later in the day. Account creation and event creation work, the same state we were in like in class.

* Login

Login page had quotes on a string I had added to test something, they have now been removed.

* Remove fontend folder

* Adds email to provider page.

* Add back Welcome message functionality to header

* Update provider.php

Provider now shows correct email address and org_name.