Standardize development environment on validated working changes (#33)
* Dockerfile: Install mysqlclient library in previous layer

This moves the `pip3 install` step for the Docker container into the
same layer as system dependencies. This way, the cached version of
mysqlclient is used whenever our project dependencies change.

Signed-off-by: Justin W. Flory <git@jwf.io>

* docker-compose/docs: Run app as single container

This commit makes changes based on in-person testing and feedback
collected from @ldidonato and @leong96 to get a working Docker Compose
environment working on multiple operating systems.

The docker-compose.yml is reverted to an earlier version. The previous
version created five versions of the application in parallel. This is
hard to test since each team member is using a specific environment to
them. Instead, this commit creates a common environment for everyone to
depend on for the app to work. Doing it this way allows us, as a group,
to quickly be aware if there is a breaking change made to the app
(because it will break for everyone instead of just one person). We
discovered some of this difficulty when debugging with Lauren, Harrison,
and I today.

Lastly, I also updated the docs for the changed port address of running
the application. I changed the port because it is an uncommon port and
it appeared macOS was already doing something on port 8000 and blocked
the app from starting. Using port 16000 for local development was a
random port pick to avoid battling with conflicting ports.

Signed-off-by: Justin W. Flory <git@jwf.io>

* docker-compose: Run app on port 8000 on localhost

This commit undoes a change in a previous commit to run the app on port
16000. Admittedly it is a non-standard port and in the case of this
morning, another container was likely already running on the port.

Docs are also updated.

cc: @nlMeminger

Signed-off-by: Justin W. Flory <git@jwf.io>