templates: Restore @ldidonato CSS, hook into static dir (#35)
This commit accomplishes two things:

1. Restores CSS changes from @ldidonato lost in The Big Merge
2. Hooks CSS changes into our base HTML template

In more detail, a site might normally ship the CSS and HTML files in the
same directory. In this case, the `templates/` directory does not truly
contain HTML, but rather HTML with the Jinja2 language embedded in,
which _then_ generates actual HTML.

In this case, Django serves things like CSS, JavaScript, and images from
a static directory. This is useful for caching and the files will get
reused across the application.

I tested this locally with docker-compose and confirmed the home page
loaded with the custom CSS included in this commit.

Signed-off-by: Justin W. Flory <git@jwf.io>