Docker: Auto-reload changes without rebuilding container (#36)
* Remove unused pieces of original template project

Signed-off-by: Justin W. Flory <git@jwf.io>

* Docker: Auto-reload changes without rebuilding container

This commit is more exciting! This makes our Docker development
experience less of a headache. Now, it is not necessary to restart the
Docker container when you make changes to the Django app. Instead,
changes made to the project will be reflected immediately in the
front-end. Just press Ctrl+R. :)

Worth noting. The only time you _will_ need to rebuild the container is
when changing project dependencies, e.g. updating the Pipfile or
updating dependencies. Otherwise you will use cached containers without
the right dependencies if you don't rebuild.

Also something I experienced, `docker-compose up --build` did not
successfully start the container for me. I literally **had** to build it
a step separately, and then run it for the Django container to be
reachable in my browser. Might be similar to @leong96's issues on Oct.
7th evening.

Signed-off-by: Justin W. Flory <git@jwf.io>