dependencies: Merge docs dependencies into project dependencies
This commit merges the two Pipenv environments in the project together.
Formerly, the docs included its own Pipfile with its own dependencies.
But... it is confusing to have two Pipfiles together in the same repo.
It's sort of like running two virtual machines together to test one
project. :thinking:

Anyways, this is part of a gradual clean-up of the docs to make them
more approachable to work on.

Signed-off-by: Justin W. Flory <git@jwf.io>