Update project dependencies to latest stable versions
This commit changed the `psycopg2` dependency to `psycopg2-binary`. This
is done because the original package name is built from source and
requires Postgres development headers to be installed in the system. I
think it makes more sense to install the pre-built binary version
suggested in the error message. This change reduces friction for someone
else working on the project later.

In doing this, I also updated project dependencies. This needs
definitive action in issue #38.

Signed-off-by: Justin W. Flory <git@jwf.io>