Add issue templates for faster filling of issues
This commit adds a set of templates for us to use for quick entry of
information into GitHub issues. It includes three templates:

* New story: front-end
* New story: back-end
* Bug report

"Stories" are the template that existed before, but I split it up by
front-end and back-end to use different default tags and assignees. A
bug report is meant to be a quick and simple template to noting a bug as
we go.

Signed-off-by: Justin W. Flory <git@jwf.io>