settings: Load DATABASE settings from YAML config (closes #12)
This commit removes the `DATABASE` options from the `development.py` and
`production.py` files and moves it into the `base.py` file. Instead of
hard-coding them into the application, the config values are retrieved
from a YAML config file at the root of the project directory.

Because the `config.yml` path is ignored, downstream users can copy the
`config.yml.example` file to `config.yml` and edit it freely as needed,
no matter what environment you are using or where your database is.

cc: @nlMeminger

Closes #12.

Signed-off-by: Justin W. Flory <git@jwf.io>