settings: Load YAML config path by CONFIGPATH env variable
This commit adds an option to specify a path to a config file as an
environment variable. If an environment variable is not set, the app
falls back on the `config.yml` at the root of the project directory.

This allows us some flexibility in how we end up deploying this in the
future and how we write our documentation.

Signed-off-by: Justin W. Flory <git@jwf.io>