ci: Add Travis job for docs and building Docker image
This commit changes our Travis CI build settings. It creates a job
matrix to do two simultaneous tests to ensure things are working:

* Build and generate our docs successfully
* Build (but not run) the Docker container successfully

This makes our tests more meaningful so we always know that the Docker
image is building successfully in `master` branch. This makes it easier
to debug in local environments for build errors.

Signed-off-by: Justin W. Flory <git@jwf.io>