Commit (#69)
The following commit makes it so that if multiple events are occuring at the same coordinates, they will appear in the same map marker. Originally, the event that was last in the database would be the only one to appear if the above scenario happened. To accomodate, much of the map marker functionality (building a string with the marker text) has been moved to views.py.

In addition, some print statements in views.py that were around for debugging have been removed.