Commits (#87)
* All Admins / Database

This commit now lists the admins in the all admins page, and removes the admins from the all providers page.

In addition, the database calls to create new admin users is in place.

* Change Name

This commit implements the Change Name functionality. Note that at the moment, the provider page has to be refreshed again for the changes to be visible. All other pages will acknowledge the change immidiately.

* Update allAdmins.php

Set href to mailto