migrations: Update/delete migrations to fix startup bug (#96)
This commit does two things:

* Updates 0001_initial migration with new metadata changes
* Removes 0003_auto_2019… migration

The third migration file causes the app to fail on start-up because it
is searching for a 0002 migration file that doesn't exist. This file was
accidentally committed in #86.

This commit requires us to start thinking of how to use migrations
though. Going forward, we might want to start committing these files
once we are confident about the base database schema. Will revisit.

Signed-off-by: Justin W. Flory <git@jwf.io>