Docker Compose: Use named volume for MariaDB (#95)
* Docker Compose: Use named volume for MariaDB

This commit uses a named volume to work around a Windows platform bug
with @leiyinmon and possibly @eguy006's setup. Using a volume in the
same directory as the project causes containers to not start:

https://github.com/docker-library/mariadb/issues/38

Tested and confirmed locally on Linux/Fedora. Did not get a chance to
test in person with @leiyinmon.

Signed-off-by: Justin W. Flory <git@jwf.io>

* Delete 0003_auto_20191022_0148.py