Split category table write from admin account creation (#97)
* Exception Catching

Catches another exception where if the database has no tables, it will be impossible to run the server or perform migrations.

Also patched the modal's ok button to work, needed to change it's type to submit.

* Seperate database catagory table writing from the create admin account function

This commit modifies the Create Administrator Account link in the header. Whereas before, it created an admin account and wrote the categories table, this commit splits it into two different links. This appeared to be an issue during today's meeting.