Forgot Password
Models - Added in table for the reset urls.
Base.py - Email information
ChangePWLogout - A new page, mainly a copy of the change password page, with fields renamed for the new purpose. Modal removed, as it is not used in this context.
resetPW - Hook up form to back end.
URLs - add changePWLogout to list of URL's.
views.py - Add code to send an email.