More email implementation
Functions - moved send_email off into a function in a seperate file.
Views - Added the approprate imports to get the function file. Then removed the send_mail function off here to become send_email in the functions file. Also fixed a bug in the reset process, a single quote was causing the reset string to not match. To demo, the send email function was also added to the login process. The user will get a "You logged in!" email upon login.