More email implementation (#121)
* More email implementation

Functions - moved send_email off into a function in a seperate file.
Views - Added the approprate imports to get the function file. Then removed the send_mail function off here to become send_email in the functions file. Also fixed a bug in the reset process, a single quote was causing the reset string to not match. To demo, the send email function was also added to the login process. The user will get a "You logged in!" email upon login.

* Adds timestamp for last login column

Models - add last login column.
All Admins - added the last_login value to the for loop.
Views - Added ability to stick a timestamp in. Removed last 7 characters of the result to omit milliseconds.

* Update 0001_initial.py

This contains the latest migrations.

* admin@admin.admin -> grasatest@yahoo.com

Header - removed the activate administrator link, the bug was resolved.
Views - Replaced the admin@admin.admin address with grasatest@yahoo.com, which is an actual email address.

* Moved category writing function.

Moved the commands that wrote the categories table to the functions.py file. Removed them from views.py

* Update views.py