Update views.py
Earlier commits broke the situation where a provider could type the admin.php url and get redirected to the index page. This has now been switched as per a suggesion in issue #51 so now typing admin.php when a provider witll go to the provider page, and vice-versa for admins trying to access provider.php.