docs: Add developer troubleshooting FAQ
This commit creates a new docs page to collect troubleshooting tips and
tricks on the application. I started it with some notes that @leong96
sent me on debugging database change issues.

Ideally, this is the go-to place we can put things that affect all of us
and can speed up our development time.

Signed-off-by: Justin W. Flory <git@jwf.io>