Email functionality / Username change
This commit implements the email functionality requested in Slack, as well as implements the username change feature.

functions.py: Moved the change username function to functions.py, saves us a couple lines of code in views.

admin.php: hooked up the change username form to the backend.

allAdmins: Hooked up the create admin form to the backend.

allusers: Hooked up the invite provider form to the backend.

provider: hooked up the change email form to the backend.

views: Added all calls to send_email to send emails in the situations specified in Slack.