Remove Django User
This commit adds on to the delete user feature; as it will now also delete the Django User, in addition to removing the database entry.

Additionally, admin users are now also made superusers to log into the django admin portal. If a person were to create a superuser via createsuperuer, it would mess up the ID numbering in the table.