'All' Pages all done + bug fixes (#148)
* 'All' Pages all done + bug fixes

-Added duplicate entry checks and feedback for add admin
- Removed isActive, non-specific gender,  not-provided, and distance filter
- Made delete in allUsers work, deletes user and all their events, then rebuilds index
-Hooked up allEvents page and set delete event to rebuild index
-Added security to all 'all' pages
-Added number of events to all providers
- Fixed bug where providers could not delete their events from provider.php
- Fixed bug where changing name in provider would cause events to not show up until reload
- Fixed bug where cancelling an event edit would cause a page break
- Fixed bug where fees would appear to 1 decimal point instead of 2

* Remove Django User

This commit adds on to the delete user feature; as it will now also delete the Django User, in addition to removing the database entry.

Additionally, admin users are now also made superusers to log into the django admin portal. If a person were to create a superuser via createsuperuer, it would mess up the ID numbering in the table.