Sync user_id with id.
This patches up an issue regarding logins, occasionally the id and user_id of the tables may go out of sync for one reason or another, preventing a login from successfully happening. The try-catch block checks that the user_id (which should always be accurate) matches the userinfo table's primary key, if not, the PK gets altered to become the user_id.

This happens at registration so that the new PK isn't taken by another user. It also is the case so that id conflicts are avoided when events are submitted.