URL fix (#158)
* Invalid Email Format Handling

This commit alters the allAdmin page and view so that if the inputted email is not in a typically valid format, the account will still be made, but the user will be notified that the email was not sent..

* Fix URL

This fixes an issue in the event page where the event URL goes to the "event locator website/url". It now uses two "/"'s to specify that this is not a relative path.

Seperately, the print to console for the SMTP exception has been removed.