config: Add `admin_email` to example config
This commit adds the `admin_email` field to the example config.
Originally I was going to load this into `base.py`, but since the
`config` object is already read into the base settings, it should be
accessible in the application, in the same way the email settings are
with @leong96's SMTP server settings.

When @eguy006 has a chance to work on the initialization / start-up
piece, the email is available in the environment when he needs it. :+1:

Related to #162.

Signed-off-by: Justin W. Flory <git@jwf.io>