dependencies: Add black to development dependencies
This commit adds Black, a Python code formatter, to our development
dependencies in the Pipfile:

    https://github.com/psf/black

This is a useful tool for inspecting code quality and improving
readability. This commit is adding it to the dependencies, even though
nothing is actually using it just yet. I'll probably spend more time on
this in the second half of Sprint 5.

Signed-off-by: Justin W. Flory <git@jwf.io>