Switch PHP extensions to HTML (closes #156)
This commit is a large commit, but it refactors our files and code to
change any existing PHP file extension to an HTML extension. This
required a few changes, but here's the process I tried to follow:

1. Deleted `index.php` - seems like a duplicate of `search/search.html`?
2. If possible, wrap a page up into a `reverse()` method.
3. Change hard-coded pointers to PHP pages to updated HTML path

This could use testing and I left comments in some areas of the
front-end code that I wasn't sure about.

Closes #156.

Signed-off-by: Justin W. Flory <git@jwf.io>

Update views.py

Minor formatting changes that allow the app to start up again.

Closed a parenthesis in def admin, removed a line break in the invite message.

Signed-off-by: Harrison Leong