views: Use reverse() for homepage redirects (#168)
This commit changes our existing redirects in our application to make
consistent use of the `reverse()` method. This uses the `name` metadata
set in the `urls.py` file. When you call something like
`reverse('search')`, this instructs Django to return the correct URL
string for whatever page matching the `search` name in the `urls.py` is.

The benefit of doing this is we don't have to remember where to update
things like file names (part of issue #156). So whenever we do switch
over the file extensions, we won't have to update multiple places in our
code when we do.

Contributes to #156 but does not close it.

Signed-off-by: Justin W. Flory <git@jwf.io>