Update views.py
This patches a function of the Create New Provider button, which advances the user_id counter, even if the admin user was not created. It now checks that an admin user exists first. If so, the process is aborted and the user is sent to the search page. If there are no admins, the admin is created like normal.