Fix various misc. bugs with URLs and paths
Sadly the `reverse('page_name')` prefixed a slash character to the
string, so the following would evaluate as so:

    reverse('login_page') == /login.html

Unfortunately, the slash throws off everything. I manually confirmed
this by hard-coding the string with and without the slash. Unfortunately
there is not enough time right now to dig into the root cause.

Signed-off-by: Justin W. Flory <git@jwf.io>