Switch PHP extensions to HTML (closes #156) (#169)
* Switch PHP extensions to HTML (closes #156)

This commit is a large commit, but it refactors our files and code to
change any existing PHP file extension to an HTML extension. This
required a few changes, but here's the process I tried to follow:

1. Deleted `index.php` - seems like a duplicate of `search/search.html`?
2. If possible, wrap a page up into a `reverse()` method.
3. Change hard-coded pointers to PHP pages to updated HTML path

This could use testing and I left comments in some areas of the
front-end code that I wasn't sure about.

Closes #156.

Signed-off-by: Justin W. Flory <git@jwf.io>

Update views.py

Minor formatting changes that allow the app to start up again.

Closed a parenthesis in def admin, removed a line break in the invite message.

Signed-off-by: Harrison Leong

* misc: Remove unused apps/base cruft

This commit removes a large amount of code that is not being used in the
project. This was originally related to how the template we started with
was set up, but we ended up not using that structure.

Related to #123.

Signed-off-by: Justin W. Flory <git@jwf.io>

* settings: Consolidate duplicates, reformat with Black

Related to #123.

This commit mostly consolidates duplicates across our three config
files. There were identical config settings repeated in `development.py`
and `production.py` from `base.py`, so where there were duplicates in
the child settings classes, I removed them.

Additionally, this commit also reformats our settings files with the
Black code linter, which is part of the work on cleaning up the code
with #123 and #164.

Signed-off-by: Justin W. Flory <git@jwf.io>

* Fix various misc. bugs with URLs and paths

Sadly the `reverse('page_name')` prefixed a slash character to the
string, so the following would evaluate as so:

    reverse('login_page') == /login.html

Unfortunately, the slash throws off everything. I manually confirmed
this by hard-coding the string with and without the slash. Unfortunately
there is not enough time right now to dig into the root cause.

Signed-off-by: Justin W. Flory <git@jwf.io>