Add GitHub Action for Black code formatter
This commit adds a new GitHub Action for reformatting Python code
according to the requirements / specs of PEP-8 / Black. This is part of
issue #123 and unblocks PR #164.

See links below for more information:

    https://github.com/psf/black
    https://github.com/lgeiger/black-action

The expected outcome of this commit is that new pull requests should
automatically have their changes reformatted by Black once they are
opened. This ensures that everyone on the team writes code in the same
style.

This is not yet tested.

Signed-off-by: Justin W. Flory <git@jwf.io>