Add SMTP config options, change template, refactor functions.py (#179)
* config: Add SMTP server settings as option

Part 1 of 5 for #157.

This commit adds `outgoing_smtp` as a config dict for collecting details
for an outgoing email server. The config values are passed into the
Django base settings for use across the application.

Signed-off-by: Justin W. Flory <git@jwf.io>

* templates: Rename aboutContact.html -> about.html

This commit renames the page name to make a more SEO-friendly and
shorter resolveable URL.

Signed-off-by: Justin W. Flory <git@jwf.io>

* Refactor: Refactor functions.py with Black (part of #123)

This commit cleans up `functions.py` to follow Black formatting rules.
This is part of the work for #123. This is a precursor commit to
organizational contact info.

Signed-off-by: Justin W. Flory <git@jwf.io>