Reset PW Expiration (#186)
* Reset PW Expiration

This commit makes several changes to the Reset Password Functionality. Originally, a reset_string would stay in the table until it's associated account's password was reset. This has been altered.

Instead, a timestamp for 10 minutes past the strings' creation is now a part of the table entry. If a user visits the reset password form URL after that timestamp, the site will instead show a "Link Expired" message, the actual form will not appear, and the table entry will remove itself. This will also occur if the reset string is not in the table (maybe due to trying to use the URL twice).

If a password was set correctly, then the table entry will continue to delete itself like before so the URL cannot be reused successfully.

In addition, another message is now in place if the inputted passwords (new and confirm) do not match.

Models: Add field for the expiration time.
resetPWForm: Add messages for if the link expired, and if the passwords do not match.
Views: Add amount of time till link expiration in the resetPW email, added code and comments for code that handles the resetPW process and the reset_string.

* Update views.py

Change 10 minutes to 60.

* Update views.py

Amended comment that indicated 10 minutes to 60.

* Reset PW Expiration

This commit makes several changes to the Reset Password Functionality. Originally, a reset_string would stay in the table until it's associated account's password was reset. This has been altered.

Instead, a timestamp for 10 minutes past the strings' creation is now a part of the table entry. If a user visits the reset password form URL after that timestamp, the site will instead show a "Link Expired" message, the actual form will not appear, and the table entry will remove itself. This will also occur if the reset string is not in the table (maybe due to trying to use the URL twice).

If a password was set correctly, then the table entry will continue to delete itself like before so the URL cannot be reused successfully.

In addition, another message is now in place if the inputted passwords (new and confirm) do not match.

Models: Add field for the expiration time.
resetPWForm: Add messages for if the link expired, and if the passwords do not match.
Views: Add amount of time till link expiration in the resetPW email, added code and comments for code that handles the resetPW process and the reset_string.

* Update views.py

Change 10 minutes to 60.

* Update views.py

Amended comment that indicated 10 minutes to 60.

* Update views.py

Swapped some SQL queries for the appropraite django commands.