config: Load organization contact info from config (closes #157)
This commit does a few things, but its primary goal is to load info
about an organization (e.g. business hours, address, social media, etc.)
from the standard config file. This is done with changes to `views.py`,
`about.html`, and `config.yml.example`.

This branch was old and I was also working on other changes that got
mashed up with this commit. Some of them are reordering package import
statements according to PEP-8, others are whitespace changes… I know it
doesn't make it the easiest commit to review, I'll work to avoid this in
the future.

I successfully tested this changes on my own, but extra validation is
welcome.

**NOTE**: This requires changes to the `config.yml` file we use. For
local development, I suggest copying `config.yml.example` and
overwriting the `config.yml`, as the example file is written to work
out-of-the-box with local development.

Closes #157. There are no more planned additions to the config file
beyond this point. Initial admin account will be carried out in #162.

Signed-off-by: Justin W. Flory <git@jwf.io>