Add CODEOWNERS GitHub config file (#199)
This commit adds a CODEOWNERS file to the GitHub repository setting.
This change will be used for two purposes:

* When certain files are changed in a PR, specific team members will be
  requested for review as subject-matter experts.
* Pull requests will be blocked from merging unless a Code Owner
  approves the PR.

I'm making this change to encourage us to be more careful with added
code now that the v1.0.0 beta is released. Everything being merged into
the project, as much as possible, should have a clear focus. Peer
reviews are an extra step we can take to prevent bugs before they are
committed; it is important we use good code reviews in our workflow.

Signed-off-by: Justin W. Flory <git@jwf.io>