Generate all send_email() messages from templates (part of #134)
This commit converts all send_email() function calls to use a set of
templates for the subject lines and mail bodies. All templates are found
in `grasa_event_locator/templates/messaging/`.

The benefit of this change is it becomes easier to change the content of
email messages without changing code. Additionally, it also allows the
potential for us to use pretty HTML documents to send prettier emails,
if we have enough time to get on top of that.

Part of #134, but does not close it yet. We still need emails altogether
for user account denial.

Signed-off-by: Justin W. Flory <git@jwf.io>

docs: Add README to messaging folder with guidance

This commit adds a README to the `templates/messaging` directory to
explain what the files are and typically used for. Most importantly, it
makes a note about the newline character when editing these files. Some
editors are automatically configured to add this character to the end of
a file, so when editing these messaging subject templates, one must use
caution and know their editor.

Signed-off-by: Justin W. Flory <git@jwf.io>