settings: Add site_name to set project hostname
This commit adds a new setting to the config file, `site_name`. This
variable will be used across our application to generate links to other
places in the application. It is important this link is always correct
and configured because it will be used extensively in future PRs.

All team members should update their `config.yml` from the example once
this PR merges.

Part of #157, but does not close it.

Signed-off-by: Justin W. Flory <git@jwf.io>