Add Django Debug Toolbar to dev environment site
This commit is cool! It adds a handy debug toolbar for use in the local
development environment. This gives us some extra tools to inspect what
is happening on a page, e.g. SQL queries, static files, templates, and
more.

The cool part is, this will only show up when doing local development.
On the production site, a different settings file is used, and the debug
toolbar is not imported.

I hope this helps us for the Bonus Sprint QA work we are planning soon.

Signed-off-by: Justin W. Flory <git@jwf.io>