Complete pop-ups
This completes the modal functionality for sending messages to the user.

admin.html: Linked forms to the back end.
edit_denied_confirm_mail.txt: message altered to include the denied reason.
provider_account_denied_mail.txt: message altered to include the denied reason.
views.py: Re-imported some classes that were removed previously but are still used in views. Implemented code in the deny functions and the admin view to pass the reason to the context of the email.

Removed comment regarding needing to insert a wrong password message. It's already there :D