:beetle: activities: Expand STEM acronym fully
I changed this because I believe not every user will always understand
what S.T.E.M. stands for. I expanded it in this commit so it is always
explicitly stated.

This fixes a bug I introduced in #200 because I renamed the category in
the database during initialization (so in the back-end, this change
already exists). I discovered they were wrong when I tried to create an
event in the front-end with the STEM category and it broke.

Signed-off-by: Justin W. Flory <git@jwf.io>