config/site_name: Use hostname from config in emails
This commit changes the site_name dict around in the config file. It
adds two fields: `prefix` and `fqdn`. These are in turn used in the
email template files to always use the configured URLs in places like
email.

There is probably a more elegant way to do this. But I am satisfied
enough with this because it puts the hostname in control of the user,
and so long as we are consistent about pulling the hostname in this way
across the back-end, it poses low risk.

Technically, this is lagging work on #157.

Signed-off-by: Justin W. Flory <git@jwf.io>