:lock: config: Rework HTTP/HTTPS as a boolean
This addresses @nlMeminger's feedback in #213 and I think is actually a
more elegant way of doing this. Basically, there is a hidden string
inside of the emails now that checks if HTTPS is set to true. If it is,
it includes an 's' character. Otherwise, Jinja silently ignores it when
writing out the HTML.

Signed-off-by: Justin W. Flory <git@jwf.io>