:memo: docs/edit-event: Explain risk of multiple edits
Addresses @leong96's feedback in #214.

This commit explains why creating multiple edits is not suggested, and
that the newest edit is always the final version (i.e. changes are not
preserved across edits).

Admittedly, this could be more intuitive but we need more time than we
have to polish this up. But I think the documentation provided here
reduces the risk of the software reaching an unexpected state, provided
the documentation is provided and used as a resource by end-users.

Part of #191.

Signed-off-by: Justin W. Flory <git@jwf.io>