:gear: config: Add sysadmin_email for logging
This commit adds another option to our config file to set the admin
email for receiving automatic error notifications. This could also be
useful for debugging in some instances.

If I have time, I want to work on logging functionality across the app
and this would hook into that later. But for now, this sets it
internally to the Django settings.

Part of #123.

Signed-off-by: Justin W. Flory <git@jwf.io>