:memo: docs: Change metadata used for PDF export (#218)
This commit changes some lines I didn't realize were important for the
PDF generation of our documentation. This makes sure that Team Platypus
is attributed for the project, not just me. :tongue:

This needs to be merged ASAP so I can generate a final PDF copy of our
documentation to submit by the dropbox deadline.

Signed-off-by: Justin W. Flory <git@jwf.io>