:sparkles: Use admin contact email in Alt. Contact modal (closes #222)
This commit closes #222.

This makes some changes to the back-end code to make sure that the
config file is passed into the context of the rendering engine. The
front-end now has the Expressions in place to parse the admin email
where it should be in the Alternative Contact "more info" modal.

Thanks @ldidonato for catching this.

Signed-off-by: Justin W. Flory <git@jwf.io>