:memo: docs: Refactor Admin Quickstart
This commit is mostly layout, structure, and formatting changes to the
Admin Quickstart page to bring it in line with our documentation.
Notably, I also opted in this commit to NOT write a dedicated page to
explain the config file, but instead, I link to the example config file
itself since the file is well-documented in-line. I don't want to repeat
myself in two different places.

Part of #122.

Signed-off-by: Justin W. Flory <git@jwf.io>