Admin self-delete  (#235)
* Update allAdmins.html

Alters the delete button in allAdmins, so the currently logged-in admin will see a "1" in their delete button table entry. This of course, will need to get replaced.

* Current User

Switch to Current User

* Update allAdmins.html

Alters the delete button in allAdmins, so the currently logged-in admin will see a "1" in their delete button table entry. This of course, will need to get replaced.